Pull Layouts et al. out of Show.kt.
