+++ src/commonMain/kotlin/baaahs/show/mutable/MutableLayouts.kt
@@ -1,0 +1,20 @@
+package baaahs.show.mutable
+
+import baaahs.show.Layouts
+
+class MutableLayouts(baseLayouts: Layouts) {
+    var panelNames = baseLayouts.panelNames.toMutableList()
+    val map = baseLayouts.map.toMutableMap()
+
+    fun copyFrom(layouts: Layouts) {
+        panelNames.clear()
+        panelNames.addAll(layouts.panelNames)
+
+        map.clear()
+        map.putAll(layouts.map)
+    }
+
+    fun build(): Layouts {
+        return Layouts(panelNames, map)
+    }
+}
