+++ src/commonMain/kotlin/baaahs/show/mutable/MutableShow.kt
@@ -244,23 +244,6 @@ class MutableShow(
     }
 }
 
-class MutableLayouts(baseLayouts: Layouts) {
-    var panelNames = baseLayouts.panelNames.toMutableList()
-    val map = baseLayouts.map.toMutableMap()
-
-    fun copyFrom(layouts: Layouts) {
-        panelNames.clear()
-        panelNames.addAll(layouts.panelNames)
-
-        map.clear()
-        map.putAll(layouts.map)
-    }
-
-    fun build(): Layouts {
-        return Layouts(panelNames, map)
-    }
-}
-
 class MutablePatch {
     val id: String = randomId("patch-editor")
